Renames the typescript tests folder to TypeScriptBetaTests
