+++ TypeScriptBetaTests/TypeScriptBetaTests.csproj
